chore: relocate Model3D component to celestial directory and update import path
