+++ src/components/celestial/Model3D.tsx
